refactor: moved services and utils to composables
