+++ src/composables/services/index.ts
@@ -1,0 +1,1 @@
+export { default as useMoviesService } from './movies'
