+++ src/composables/services/movies.ts
@@ -1,9 +1,9 @@
 import type { apiMovieType } from '@/types/types'
-import { useUrlHandler } from '@/utils/useUrlHandler'
+import { useUrlHandler } from '@/composables/utils'
 
 const params = '?include_adult=false&include_video=false&language=en-US&page=1&sort_by=popularity.desc'
 
-export function useMoviesService() {
+export default function useMoviesService() {
     const { headers, buildUrl } = useUrlHandler()
 
     // methods
